+++ src/cot_assert/_runtime.py
@@ -3,9 +3,10 @@
 from __future__ import absolute_import, division, print_function
 
 from ._error import AssertSite, annotated
+from ._values import message as m
 from ._values import value as v
 
-__all__ = ["AssertSite", "fail", "v"]
+__all__ = ["AssertSite", "fail", "m", "v"]
 
 
 def fail(site, path, values, msg):